merge

+++ trunk/iPhone/Sample/Sources/ZipArchive/ZipArchive.mm
@@ -157,7 +157,7 @@
 		unz_global_info  globalInfo = {0};
 		if( unzGetGlobalInfo(_unzFile, &globalInfo )==UNZ_OK )
 		{
-			NSLog(@"%d entries in the zip file", globalInfo.number_entry);
+			NSLog(@"%u entries in the zip file", (uint)globalInfo.number_entry);
 		}
 	}
 	return _unzFile!=NULL;
